requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.0.6
django-cleanup==8.1.0
django-ninja==1.3.0
django-widget-tweaks==1.5.0
djlint==1.34.1
EditorConfig==0.12.4
frozenlist==1.4.1
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
jsbeautifier==1.15.1
json5==0.9.25
multidict==6.0.5
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2023.12.25
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.5
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
